+++ src/main/java/br/com/estudigi/api/model/Test.java
@@ -16,7 +16,7 @@ public class Test {
 
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
-    Long id;
+    Integer id;
 
     String name;
     Subject subject;
