+++ src/main/java/br/com/estudigi/api/repository/ChoiceRepository.java
@@ -5,5 +5,5 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface ChoiceRepository extends JpaRepository<Choice, Long> {
+public interface ChoiceRepository extends JpaRepository<Choice, Integer> {
 }
